pyproject.toml
[tool.poetry]
name = "module-1"
version = "0.1.0"
description = "Web scraping application for books.toscrape.com"
authors = ["Johnz1n <joaopedroborca@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
